Sync feature status, diff and merge files to home
Added the main feature of the sync_with_homedir.sh script; file synchronization betweeen this repo and the home directory.
The user can choose to only view the sync status, or a diff, or merge.

The script will not start in merge-mode, it will show the help menu.

The merge mode will check if the file exists already, if i does it will do a vim diff of the files isntead of just writing over.